requirements.txt
pandas
nltk
gensim
scikit-learn
transformers
torch
umap-learn
matplotlib
seaborn
hdbscan
